pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ln2t_watchdog"
version = "0.1.0"
description = "Automated nightly data-processing scheduler for ln2t_tools pipelines"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
ln2t-watchdog = "ln2t_watchdog.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
